Revert "935 trend legend (#3434)"
This reverts commit cd660e52f8e3bc77e28ce0ec9e76849224883585.
