+++ frontend/src/lib/utils.tsx
@@ -241,6 +241,8 @@ export function formatLabel(
         label += ` (Active Users) `
     } else if (['sum', 'avg', 'min', 'max', 'median', 'p90', 'p95', 'p99'].includes(action.math)) {
         label += ` (${action.math} of ${action.math_property}) `
+    } else {
+        label += ' (Total) '
     }
     if (action?.properties?.length) {
         label += ` (${action.properties
