+++ frontend/src/scenes/insights/Insights.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react'
-import { useActions, useMountedLogic, useValues, BindLogic } from 'kea'
+import { useActions, useMountedLogic, useValues } from 'kea'
 
 import { Loading } from 'lib/utils'
 import { SaveToDashboard } from 'lib/components/SaveToDashboard/SaveToDashboard'
@@ -13,7 +13,6 @@ import { ChartFilter } from 'lib/components/ChartFilter'
 import { Tabs, Row, Col, Card, Button } from 'antd'
 import {
     ACTIONS_LINE_GRAPH_LINEAR,
-    ACTIONS_LINE_GRAPH_CUMULATIVE,
     ACTIONS_TABLE,
     ACTIONS_PIE_CHART,
     ACTIONS_BAR_CHART_VALUE,
@@ -42,9 +41,7 @@ import './Insights.scss'
 import { ErrorMessage, TimeOut } from './EmptyStates'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { People } from 'scenes/funnels/People'
-import { TrendLegend } from './TrendLegend'
 import { TrendInsight } from 'scenes/trends/Trends'
-import { trendsLogic } from 'scenes/trends/trendsLogic'
 
 const { TabPane } = Tabs
 
@@ -307,19 +304,6 @@ function _Insights() {
                                 allFilters.display === FUNNEL_VIZ && (
                                     <Card>
                                         <FunnelPeople />
-                                    </Card>
-                                )}
-                            {(!allFilters.display ||
-                                allFilters.display === ACTIONS_LINE_GRAPH_LINEAR ||
-                                allFilters.display === ACTIONS_LINE_GRAPH_CUMULATIVE) &&
-                                (activeView === ViewType.TRENDS || activeView === ViewType.SESSIONS) && (
-                                    <Card>
-                                        <BindLogic
-                                            logic={trendsLogic}
-                                            props={{ dashboardItemId: null, view: activeView }}
-                                        >
-                                            <TrendLegend />
-                                        </BindLogic>
                                     </Card>
                                 )}
                         </Col>
