+++ frontend/src/scenes/trends/viz/ActionsLineGraph.js
@@ -21,22 +21,21 @@ export function ActionsLineGraph({
         filters: filtersParam,
         cachedResults,
     })
-    const { filters, indexedResults, resultsLoading, visibilityMap } = useValues(logic)
+    const { filters, results, resultsLoading } = useValues(logic)
     const { loadPeople } = useActions(logic)
 
     const { people_action, people_day, ...otherFilters } = filters // eslint-disable-line
     const [{ fromItem }] = useState(router.values.hashParams)
 
-    return indexedResults && !resultsLoading ? (
-        indexedResults.reduce((total, item) => total + item.count, 0) !== 0 ? (
+    return results && !resultsLoading ? (
+        results.reduce((total, item) => total + item.count, 0) !== 0 ? (
             <LineGraph
                 pageKey={'trends-annotations'}
                 data-attr="trend-line-graph"
                 type={filters.shown_as === LIFECYCLE || filters.display === ACTIONS_BAR_CHART ? 'bar' : 'line'}
                 color={color}
-                datasets={indexedResults}
-                visibilityMap={visibilityMap}
-                labels={(indexedResults[0] && indexedResults[0].labels) || []}
+                datasets={results}
+                labels={(results[0] && results[0].labels) || []}
                 isInProgress={!filters.date_to}
                 dashboardItemId={dashboardItemId || fromItem}
                 inSharedMode={inSharedMode}
